Fetch data from missions endpoint
