+++ src/redux/configureStore.js
@@ -3,9 +3,11 @@ import { combineReducers } from 'redux';
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
 import rocketReducer from './rockets/rocketsSlice';
+import missionReducer from './missions/missionsSlice';
 
 const rootReducer = combineReducers({
   rockets: rocketReducer,
+  missions: missionReducer,
 });
 
 const store = configureStore({
