+++ src/services/missionService.js
@@ -1,0 +1,9 @@
+import http from '../http-common';
+
+const fetchMissions = () => http.get('/missions');
+
+const missionsService = {
+  fetchMissions,
+};
+
+export default missionsService;
